techdebt: [CHAOS-295]: Extract chaos client to a new module
