+++ 120-ng-manager/BUILD.bazel
@@ -56,6 +56,7 @@ shared_dependencies = [
     "//960-watcher:module",
     "//965-api-key-filter:module",
     "//clients/access-control/admin-client:module",
+    "//clients/chaos:module",
     "//clients/ngsettings:module",
     "//product/ci/scm/proto:ciscm_java_proto",
     "@maven//:net_rcarz_jira_client",
