+++ clients/chaos/BUILD.bazel
@@ -1,0 +1,22 @@
+load("//:tools/bazel/harness.bzl", "java_library")
+load("//:tools/bazel/GenTestRules.bzl", "run_tests")
+load("//:tools/bazel/macros.bzl", "run_analysis")
+
+java_library(
+    name = "module",
+    srcs = glob(["src/main/java/**/*.java"]),
+    visibility = ["//visibility:public"],
+    deps = [
+        "//:lombok",
+        "//970-ng-commons:module",
+        "//980-commons:module",
+        "//999-annotations:module",
+        "@maven//:com_google_code_findbugs_annotations",
+        "@maven//:com_google_guava_guava",
+        "@maven//:com_google_inject_guice",
+        "@maven//:com_squareup_retrofit2_retrofit",
+        "@maven//:org_slf4j_slf4j_api",
+    ],
+)
+
+run_analysis()
