+++ clients/chaos/src/main/java/io/harness/chaos/client/beans/ChaosApplyManifestResponse.java
@@ -5,7 +5,7 @@
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package io.harness.ng.chaos.client;
+package io.harness.chaos.client.beans;
 
 import lombok.Builder;
 import lombok.Value;
