+++ clients/chaos/src/main/java/io/harness/chaos/client/remote/ChaosHttpClientFactory.java
@@ -5,7 +5,7 @@
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package io.harness.ng.chaos.client;
+package io.harness.chaos.client.remote;
 
 import static io.harness.annotations.dev.HarnessTeam.PL;
